requirements.txt
Django==1.4.1
PIL==1.1.7
django-storages==1.1.5
boto==2.6.0
